+++ src/OAuth.php
@@ -2,9 +2,8 @@
 declare(strict_types=1);
 namespace Patreon;
 
-use ParagonIE\Certainty\RemoteFetch;
-use ParagonIE\HiddenString\HiddenString;
 use Patreon\Exceptions\CurlException;
+use SensitiveParameter;
 
 /**
  * Class OAuth
@@ -12,52 +11,17 @@ use Patreon\Exceptions\CurlException;
  */
 class OAuth
 {
-    /** @var HiddenString $client_id */
-    private $client_id;
-
-    /** @var HiddenString $client_secret */
-    private $client_secret;
-
-    /** @var string $caCertDir */
-    protected $caCertDir;
+    private string $client_id;
+    private string $client_secret;
 
     /**
      * OAuth constructor.
      *
-     * @param string|HiddenString $client_id
-     * @param string|HiddenString $client_secret
      */
-    public function __construct($client_id, $client_secret, $cacert_dir = '')
+    public function __construct(#[SensitiveParameter]string $client_id, #[SensitiveParameter] string $client_secret)
     {
-        if (!($client_id instanceof HiddenString)) {
-            $client_id = new HiddenString($client_id);
-        }
-        if (!($client_secret instanceof HiddenString)) {
-            $client_secret = new HiddenString($client_secret);
-        }
         $this->client_id = $client_id;
         $this->client_secret = $client_secret;
-        if (!empty($cacert_dir)) {
-            $this->caCertDir = $cacert_dir;
-        }
-    }
-
-    /**
-     * @return string
-     */
-    public function getCaBundleDir(): string
-    {
-        return $this->caCertDir;
-    }
-
-    /**
-     * @param string $path
-     * @return self
-     */
-    public function setCaBundleDir(string $path): self
-    {
-        $this->caCertDir = $path;
-        return $this;
     }
 
     /**
@@ -66,13 +30,13 @@ class OAuth
      * @return array
      * @throws CurlException
      */
-    public function get_tokens(string $code, string $redirect_uri)
+    public function get_tokens(string $code, string $redirect_uri): array
     {
         return $this->__update_token([
             "grant_type" => "authorization_code",
             "code" => $code,
-            "client_id" => $this->client_id->getString(),
-            "client_secret" => $this->client_secret->getString(),
+            "client_id" => $this->client_id,
+            "client_secret" => $this->client_secret,
             "redirect_uri" => $redirect_uri
         ]);
     }
@@ -82,13 +46,13 @@ class OAuth
      * @return array
      * @throws CurlException
      */
-    public function refresh_token(string $refresh_token)
+    public function refresh_token(#[SensitiveParameter]string $refresh_token): array
     {
         return $this->__update_token([
             "grant_type" => "refresh_token",
             "refresh_token" => $refresh_token,
-            "client_id" => $this->client_id->getString(),
-            "client_secret" => $this->client_secret->getString()
+            "client_id" => $this->client_id,
+            "client_secret" => $this->client_secret
         ]);
     }
 
@@ -112,7 +76,6 @@ class OAuth
         if (!defined('CURLOPT_SSL_VERIFYPEER')) {
             curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, true);
         }
-        curl_setopt($ch, CURLE_SSL_CACERT, $this->getLatestCaCerts());
 
         $response = curl_exec($ch);
         if (!is_string($response)) {
@@ -120,19 +83,5 @@ class OAuth
         }
 
         return (array) json_decode($response, true);
-    }
-
-    /**
-     * Can be overridden in derived classes.
-     *
-     * @return string
-     * @throws \ParagonIE\Certainty\Exception\CertaintyException
-     * @throws \SodiumException
-     */
-    public function getLatestCaCerts(): string
-    {
-        return (new RemoteFetch($this->caCertDir))
-            ->getLatestBundle()
-            ->getFilePath();
     }
 }