+++ tests/OAuthTest.php
@@ -1,7 +1,6 @@
 <?php
 namespace Patreon\Tests;
 
-use ParagonIE\HiddenString\HiddenString;
 use Patreon\OAuth;
 use PHPUnit\Framework\TestCase;
 
@@ -20,18 +19,7 @@ class OAuthTest extends TestCase
 
         $this->assertInstanceOf(
             OAuth::class,
-            new OAuth(new HiddenString('a'), new HiddenString('b'))
-        );
-    }
-
-    public function testCaCertBundlePersist()
-    {
-        $path = realpath(__DIR__ . '/certs');
-        $oa = new OAuth(new HiddenString('a'), new HiddenString('b'), $path);
-        $file = $oa->getLatestCaCerts();
-        $this->assertSame(
-            $path,
-            realpath(dirname($file))
+            new OAuth('a', 'b')
         );
     }
 }